update camera configuration
